refactor(download): move atomic metadata writes into core crate
